requirements.txt
absl-py==0.6.1
alabaster==0.7.12
anaconda-client==1.7.1
anaconda-navigator==1.9.6
anaconda-project==0.8.2
appdirs==1.4.3
argcomplete==1.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
astropy==3.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
blinker==1.4
bokeh==1.0.2
boto==2.49.0
boto3==1.9.70
botocore==1.12.70
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
conda==4.5.12
conda-build==3.17.5
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cryptography-vectors==2.3.1
cycler==0.10.0
cymem==1.31.2
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
distributed==1.25.1
docutils==0.14
docx2txt==0.6
EbookLib==0.15
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.6
fuzzywuzzy==0.17.0
gast==0.2.0
gensim==3.5.0
gevent==1.3.7
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.13
grpcio==1.16.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==17.3.1
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
jeepney==0.4
Jinja2==2.10
jmespath==0.9.3
jsonschema==3.0.0a3
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multipledispatch==0.6.0
murmurhash==0.28.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.2.5
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.9
numpy==1.15.4
numpydoc==0.8.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.12
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
plac==0.9.6
pluggy==0.8.0
ply==3.11
preshed==1.0.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.2.2
pymongo==3.7.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyrsistent==0.14.8
PySocks==1.6.8
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-crfsuite==0.9.6
python-dateutil==2.7.5
python-docx==0.8.7
python-dotenv==0.10.0
python-Levenshtein==0.12.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
regex==2017.11.9
requests==2.21.0
requests-oauthlib==1.0.0
rope==0.10.7
ruamel-yaml==0.15.71
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.1
scikit-multilearn==0.2.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
spacy==2.0.13
SpeechRecognition==3.6.3
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder-kernels==1.1.0
SQLAlchemy==1.2.15
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textgenrnn==1.4.1
thinc==6.10.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==18.9.0
twython==3.7.0
typed-ast==1.1.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zict==0.1.3
zope.interface==4.6.0
